Se agregaron los módulos de roles y permisos, así como se agregó datatable a cada vista table.blade.php de cada módulo.

+++ resources/views/users/table.blade.php
@@ -25,28 +25,33 @@
             </thead>
             <tbody>
             @foreach($users as $user)
-                <tr>
-                    <td class="text-center">{{ $user->name }}</td>
-                    <td class="text-center">{{ $user->email }}</td>
-                    {{-- <td>{{ $user->email_verified_at }}</td> --}}
-                    {{-- <td>{{ $user->password }}</td> --}}
-                    {{-- <td>{{ $user->remember_token }}</td> --}}
-                    <td class="text-center" style="width: 120px">
-                        {!! Form::open(['route' => ['users.destroy', $user->id], 'method' => 'delete']) !!}
-                        <div class='btn-group'>
-                            <a href="{{ route('users.show', [$user->id]) }}"
-                            class='btn btn-default btn-xs'>
-                                <i class="far fa-eye"></i>
-                            </a>
-                            <a href="{{ route('users.edit', [$user->id]) }}"
-                            class='btn btn-default btn-xs'>
-                                <i class="far fa-edit"></i>
-                            </a>
-                            {!! Form::button('<i class="far fa-trash-alt"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
-                        </div>
-                        {!! Form::close() !!}
-                    </td>
-                </tr>
+                @if ($user->email != 'cande.pacheco@gmail.com' && Auth())
+                    <tr>
+                        <td class="text-center">{{ $user->name }}</td>
+                        <td class="text-center">{{ $user->email }}</td>
+                        {{-- <td>{{ $user->email_verified_at }}</td> --}}
+                        {{-- <td>{{ $user->password }}</td> --}}
+                        {{-- <td>{{ $user->remember_token }}</td> --}}
+                        <td class="text-center" style="width: 120px">
+                            {!! Form::open(['route' => ['users.destroy', $user->id], 'method' => 'delete']) !!}
+                            <div class='btn-group'>
+                                <a href="{{ route('users.show', [$user->id]) }}"
+                                class='btn btn-default btn-xs'>
+                                    <i class="far fa-eye"></i>
+                                </a>
+                                <a href="{{ route('users.edit', [$user->id]) }}"
+                                class='btn btn-default btn-xs'>
+                                    <i class="far fa-edit"></i>
+                                </a>
+                                {!! Form::button('<i class="far fa-trash-alt"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
+                            </div>
+                            {!! Form::close() !!}
+                        </td>
+                    </tr>
+                @else
+                    
+                @endif
+                
             @endforeach
             </tbody>
         </table>
